Add ability to set a game's local app data folder name
This is necessary to support Enderal, which uses %LOCALAPPDATA%\enderal.
