+++ src/gui/helpers.cpp
@@ -158,4 +158,70 @@ int CompareFilenames(const std::string& lhs, const std::string& rhs) {
   return unicodeLhs.caseCompare(unicodeRhs, U_FOLD_CASE_DEFAULT);
 #endif
 }
+
+std::filesystem::path getExecutableDirectory() {
+#ifdef _WIN32
+  // Despite its name, paths can be longer than MAX_PATH, just not by default.
+  // FIXME: Make this work with long paths.
+  std::wstring executablePathString(MAX_PATH, 0);
+
+  if (GetModuleFileName(NULL, &executablePathString[0], MAX_PATH) == 0) {
+    auto logger = getLogger();
+    if (logger) {
+      logger->error("Failed to get LOOT executable path.");
+    }
+    throw std::system_error(GetLastError(),
+                            std::system_category(),
+                            "Failed to get LOOT executable path.");
+  }
+
+  return std::filesystem::path(executablePathString).parent_path();
+#else
+  char result[PATH_MAX];
+
+  ssize_t count = readlink("/proc/self/exe", result, PATH_MAX);
+  if (count < 0) {
+    auto logger = getLogger();
+    if (logger) {
+      logger->error("Failed to get LOOT executable path.");
+    }
+    throw std::system_error(
+        count, std::system_category(), "Failed to get LOOT executable path.");
+  }
+
+  return std::filesystem::u8path(std::string(result, count)).parent_path();
+#endif
+}
+
+std::filesystem::path getLocalAppDataPath() {
+#ifdef _WIN32
+  HWND owner = 0;
+  PWSTR path;
+
+  if (SHGetKnownFolderPath(FOLDERID_LocalAppData, 0, NULL, &path) != S_OK)
+    throw std::system_error(GetLastError(),
+                            std::system_category(),
+                            "Failed to get %LOCALAPPDATA% path.");
+
+  std::filesystem::path localAppDataPath(path);
+  CoTaskMemFree(path);
+
+  return localAppDataPath;
+#else
+  // Use XDG_CONFIG_HOME environmental variable if it's available.
+  const char* xdgConfigHome = getenv("XDG_CONFIG_HOME");
+
+  if (xdgConfigHome != nullptr)
+    return std::filesystem::u8path(xdgConfigHome);
+
+  // Otherwise, use the HOME env. var. if it's available.
+  xdgConfigHome = getenv("HOME");
+
+  if (xdgConfigHome != nullptr)
+    return std::filesystem::u8path(xdgConfigHome) / ".config";
+
+  // If somehow both are missing, use the executable's directory.
+  return getExecutableDirectory();
+#endif
+}
 }
